requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-core
azure-identity
python-dotenv
markdown2
